+++ stories/components/d3-chart/docs/examples/stacked-area-chart/legend-and-pattern-fill.tsx
@@ -1,3 +1,4 @@
+import { Flex, Box } from '@semcore/base-components';
 import {
   Plot,
   XAxis,
@@ -8,7 +9,6 @@ import {
   makeDataHintsContainer,
   ChartLegend,
 } from '@semcore/d3-chart';
-import { Flex, Box } from '@semcore/flex-box';
 import { Text } from '@semcore/typography';
 import { scaleLinear } from 'd3-scale';
 import { curveCardinal } from 'd3-shape';