+++ stories/components/input-phone/docs/examples/unknown_country_and_number_format.tsx
@@ -1,5 +1,5 @@
+import { Flex } from '@semcore/base-components';
 import { ButtonLink } from '@semcore/button';
-import { Flex } from '@semcore/flex-box';
 import CloseM from '@semcore/icon/Close/m';
 import Input from '@semcore/input';
 import { Hint } from '@semcore/tooltip';