+++ stories/components/select/docs/examples/trigger_customization.tsx
@@ -1,5 +1,5 @@
+import { Flex } from '@semcore/base-components';
 import { ButtonTrigger, LinkTrigger } from '@semcore/base-trigger';
-import { Flex } from '@semcore/flex-box';
 import Select from '@semcore/select';
 import { Text } from '@semcore/typography';
 import React from 'react';