+++ stories/patterns/ux-patterns/auto-suggest/docs/examples/autosuggest_example.tsx
@@ -1,4 +1,4 @@
-import { Box } from 'intergalactic/flex-box';
+import { Box } from 'intergalactic/base-components';
 import Input from 'intergalactic/input';
 import Select from 'intergalactic/select';
 import { Text } from 'intergalactic/typography';