+++ semcore/animation/CHANGELOG.md
@@ -18,7 +18,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Deprecated
 
-- Import it from `@semcore/core`.
+- Import it from `@semcore/base-components`.
 
 ## [2.41.4] - 2025-05-13